api/requirements.txt
click==8.1.3
Flask==2.2.3
Flask-Admin==1.6.1
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyJWT==1.7.1
python-dotenv==1.0.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
uuid==1.30
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
